requirements.txt
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
anthropic==0.21.3
schedule==1.2.1
